requirements.txt
streamlit==1.32.0
langchain==0.1.13
langchain-google-genai==1.0.1
faiss-cpu
PyPDF2==3.0.1
spacy==3.7.4
python-dotenv==1.0.1